+++ packages/documentation/src/shared/packages.data.ts
@@ -5,6 +5,8 @@ import metaComponentsAngular from '@/stories/getting-started/packages/components
 import metaInternetHeader from '@/stories/getting-started/packages/internet-header/internet-header.stories';
 import metaIntranetHeader from '@/stories/getting-started/packages/intranet-header/intranet-header.stories';
 import metaIcons from '@/stories/getting-started/packages/icons/package-icons.stories';
+import metaThemeAGGrid from '@/stories/getting-started/packages/theme-ag-grid/theme-ag-grid.stories';
+
 import { PackageType } from '@/../types';
 
 interface IPackage {
@@ -144,6 +146,26 @@ export const packages: IPackage[] = [
       alt: '',
     },
     version: `v${getVersion(DEPENDENCIES['@swisspost/design-system-icons'])}`,
+  },
+  {
+    name: 'Theme AG Grid',
+    docsStoryId: metaThemeAGGrid.id,
+    type: PackageType.Ts,
+    link: {
+      github: {
+        href: 'https://github.com/swisspost/design-system/tree/main/packages/theme-ag-grid',
+        ariaLabel: 'Source of Theme AG Grid package',
+      },
+      docs: {
+        href: generateDocsRelativeLink(metaThemeAGGrid.id),
+        ariaLabel: 'Getting started with Theme AG Grid package',
+      },
+    },
+    img: {
+      src: '/assets/images/packages/theme-ag-grid.svg',
+      alt: '',
+    },
+    version: `v${getVersion(DEPENDENCIES['@swisspost/design-system-theme-ag-grid'])}`,
   },
 ];